refactor: auth関連のファイルをhooksディレクトリに移動した
