+++ frontend/src/hooks/providers/useAuthProvider.tsx
